requirements.txt
openai
requests
schedule
tweepy
python-telegram-bot
